+++ lib/functions/vSysMac.ts
@@ -28,12 +28,14 @@ const RGX_64_QUAD_COLON = /^[0-9A-Fa-f]{4}:[0-9A-Fa-f]{2}FF:FE[0-9A-Fa-f]{2}:[0-
 /* Special Case: MMMM-MMFF-FESS-SSSS format (48-bit address converted to a 64-bit) */
 const RGX_64_QUAD_DASH = /^[0-9A-Fa-f]{4}-[0-9A-Fa-f]{2}FF-FE[0-9A-Fa-f]{2}-[0-9A-Fa-f]{4}$/;
 
+export type MAC = string & {readonly brand: unique symbol};
+
 /**
  * Validate that a provided value is a valid MAC address
  *
  * @param {unknown} val - Value to verify
  */
-function vSysMac (val:unknown):val is string {
+function vSysMac (val:unknown):val is MAC {
     if (typeof val !== 'string') return false;
 
     return (